requirements.txt
asgiref==3.2.3
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.11.1
Automat==0.8.0
backcall==0.1.0
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
constantly==15.1.0
coverage==5.0.3
cryptography==2.8
daphne==2.4.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.2.0
django-eventstream==2.6.0
django-grip==2.0.2
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.0
gripcontrol==3.3.1
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.13.0
prompt-toolkit==2.0.10
psycopg2==2.8.3
ptyprocess==0.6.0
pubcontrol==2.4.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.2
pytest==5.1.3
python-dateutil==2.8.0
python-magic==0.4.15
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
